splits data and logic quotes
