+++ src/lot_format.php
@@ -1,0 +1,26 @@
+<?php
+function format_price($price)
+{
+    $part_length = 3;
+
+    $original = ceil($price);
+
+    $result = '';
+    $length = strlen($original);
+    $parts = ceil($length / $part_length);
+
+    for ($part = 0; $part < $parts; $part++) {
+        $part_start = $length - $part * $part_length;
+        $start = $part_start < $part_length ? 0 : $part_start - $part_length;
+
+        $rank = substr($original, $start, min($part_length, $part_start));
+
+        if ($part === 0) {
+            $result = $rank;
+        } else {
+            $result = $rank . ' ' . $result;
+        }
+    }
+
+    return $result . ' ₽';
+}
