+++ templates/main.php
@@ -1,64 +1,5 @@
 <?php
-$lots = [
-    [
-        'name' => '2014 Rossignol District Snowboard',
-        'category' => 'Доски и лыжи',
-        'price' => 10999,
-        'image' => 'img/lot-1.jpg'
-    ],
-    [
-        'name' => 'DC Ply Mens 2016/2017 Snowboard',
-        'category' => 'Доски и лыжи',
-        'price' => 159999,
-        'image' => 'img/lot-2.jpg'],
-    [
-        'name' => 'Крепления Union Contact Pro 2015 года размер L/XL',
-        'category' => 'Крепления',
-        'price' => 8000,
-        'image' => 'img/lot-3.jpg'],
-    [
-        'name' => 'Ботинки для сноуборда DC Mutiny Charocal',
-        'category' => 'Ботинки',
-        'price' => 10999,
-        'image' => 'img/lot-4.jpg'],
-    [
-        'name' => 'Куртка для сноуборда DC Mutiny Charocal',
-        'category' => 'Одежда',
-        'price' => 7500,
-        'image' => 'img/lot-5.jpg'],
-    [
-        'name' => 'Маска Oakley Canopy',
-        'category' => 'Разное',
-        'price' => 5400,
-        'image' => 'img/lot-6.jpg']
-];
-
-
-function format_price($price)
-{
-    $part_length = 3;
-
-    $original = ceil($price);
-
-    $result = '';
-    $length = strlen($original);
-    $parts = ceil($length / $part_length);
-
-    for ($part = 0; $part < $parts; $part++) {
-        $part_start = $length - $part * $part_length;
-        $start = $part_start < $part_length ? 0 : $part_start - $part_length;
-
-        $rank = substr($original, $start, min($part_length, $part_start));
-
-        if ($part === 0) {
-            $result = $rank;
-        } else {
-            $result = $rank . ' ' . $result;
-        }
-    }
-
-    return $result . ' ₽';
-}
+require_once 'src/lot_format.php';
 ?>
 <section class="promo">
     <h2 class="promo__title">Нужен стафф для катки?</h2>
@@ -80,11 +21,11 @@ function format_price($price)
         <?php foreach ($lots as $lot): ?>
             <li class="lots__item lot">
                 <div class="lot__image">
-                    <img src="<?= $lot['image'] ?>" width="350" height="260" alt="<?= htmlspecialchars($lot['name']) ?>">
+                    <img src="<?= $lot['image'] ?>" width="350" height="260" alt="<?= htmlspecialchars($lot['name'], ENT_QUOTES) ?>">
                 </div>
                 <div class="lot__info">
                     <span class="lot__category"><?= $lot['category'] ?></span>
-                    <h3 class="lot__title"><a class="text-link" href="pages/lot.html"><?= htmlspecialchars($lot['name']) ?></a>
+                    <h3 class="lot__title"><a class="text-link" href="pages/lot.html"><?= htmlspecialchars($lot['name'], ENT_QUOTES) ?></a>
                     </h3>
                     <div class="lot__state">
                         <div class="lot__rate">
